+++ soya/src/helper.js
@@ -1,8 +1,0 @@
-/**
- * @CLIENT-SERVER
- */
-export default {
-  prop(value, propName) {
-    return value && value[propName] ? value[propName] : value;
-  }
-};